- Matrices work with Cartesian coordinate system (Cartesian coordinate system now implemented at OpenGL level) - Scale now goes from 0 to 2 (2 would be 100% of the with or height) making going from physics coordinates to OpenGL is simple (divide by 1024) - Re-factored Texture to use float instead of double

+++ SketchWars/src/sketchwars/character/weapon/GrenadeWeapon.java
@@ -12,7 +12,7 @@ import sketchwars.sound.SoundPlayer;
 public class GrenadeWeapon extends AbstractWeapon {   
     private double explosionRadius;
     
-    public GrenadeWeapon(Texture texture, double width, double height, ProjectileFactory projectileFactory) {
+    public GrenadeWeapon(Texture texture, float width, float height, ProjectileFactory projectileFactory) {
         super(texture, width, height, projectileFactory);
         init();
     }
